+++ rmk-macro/src/default_config/nrf52840.rs
@@ -14,6 +14,9 @@ pub(crate) fn default_nrf52840(chip: ChipModel) -> KeyboardConfig {
             get_usb_info("nrf52840").unwrap(),
             BleConfig {
                 enabled: true,
+                // Use nice!nano's default divider config
+                adc_divider_measured: Some(2000),
+                adc_divider_total: Some(2806),
                 ..Default::default()
             },
         ),