+++ rmk/src/debounce/mod.rs
@@ -8,6 +8,8 @@ static DEBOUNCE_THRESHOLD: u16 = 10;
 
 pub(crate) trait DebouncerTrait {
     fn new() -> Self;
+
+    /// The `in_idx` `out_idx` can be used as two normal dimensions.
     fn detect_change_with_debounce(
         &mut self,
         in_idx: usize,