+++ scripts/check_all.sh
@@ -10,6 +10,7 @@ cd examples/use_rust/nrf52832_ble && cargo build --release && cd ../../..
 cd examples/use_rust/nrf52840 && cargo build --release && cd ../../..
 cd examples/use_rust/nrf52840_ble && cargo build --release && cd ../../..
 cd examples/use_rust/rp2040 && cargo build --release && cd ../../..
+cd examples/use_rust/rp2040_direct_pin && cargo build --release && cd ../../.. 
 cd examples/use_rust/stm32f1 && cargo build --release && cd ../../..
 cd examples/use_rust/stm32f4 && cargo build --release && cd ../../..
 cd examples/use_rust/stm32h7 && cargo build --release && cd ../../.. 
@@ -22,6 +23,7 @@ cd examples/use_config/nrf52832_ble && cargo build --release && cd ../../..
 cd examples/use_config/nrf52840_ble && cargo build --release && cd ../../.. 
 cd examples/use_config/nrf52840_usb && cargo build --release && cd ../../.. 
 cd examples/use_config/rp2040 && cargo build --release && cd ../../.. 
+cd examples/use_config/rp2040_direct_pin && cargo build --release && cd ../../.. 
 cd examples/use_config/stm32f1 && cargo build --release && cd ../../..
 cd examples/use_config/stm32f4 && cargo build --release && cd ../../..
 cd examples/use_config/stm32h7 && cargo build --release && cd ../../.. 
@@ -33,6 +35,7 @@ cd examples/use_rust/nrf52832_ble && cargo clean && cd ../../..
 cd examples/use_rust/nrf52840 && cargo clean && cd ../../..
 cd examples/use_rust/nrf52840_ble && cargo clean && cd ../../..
 cd examples/use_rust/rp2040 && cargo clean && cd ../../..
+cd examples/use_rust/rp2040_direct_pin && cargo clean && cd ../../.. 
 cd examples/use_rust/stm32f1 && cargo clean && cd ../../..
 cd examples/use_rust/stm32f4 && cargo clean && cd ../../..
 cd examples/use_rust/stm32h7 && cargo clean && cd ../../..  
@@ -46,6 +49,7 @@ cd examples/use_config/nrf52832_ble && cargo clean && cd ../../..
 cd examples/use_config/nrf52840_ble && cargo clean && cd ../../.. 
 cd examples/use_config/nrf52840_usb && cargo clean && cd ../../.. 
 cd examples/use_config/rp2040 && cargo clean && cd ../../.. 
+cd examples/use_config/rp2040_direct_pin && cargo clean && cd ../../.. 
 cd examples/use_config/stm32f1 && cargo clean && cd ../../..
 cd examples/use_config/stm32f4 && cargo clean && cd ../../..
 cd examples/use_config/stm32h7 && cargo clean && cd ../../..