+++ rmk/CHANGELOG.md
@@ -7,6 +7,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add layout macro to! and df!
+- Make nRF52840 voltage divider configurable
+
+### Changed
+
+- Use `User11` to manually switch between USB mode and BLE mode
+
+### Fixed
+
+- Fix nRF52840 linker scripts for nice!nano
+
+## [0.4.1] - 2024-10-31
+
+### Fixed
+
+- Fix lagging for split peripheral 
+
+### Added
+
+- Direct pin mod. Including `DirectPinMatrix`, `run_rmk_direct_pin` functions etc.
+- Added pin active level parameter `low_active` to direct pin.
+- Support no_pin for `DirectPinMatrix`.
+
 ## [0.4.0] - 2024-10-28
 
 ### Added